# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST in Golang with Gin Framework",
        "contact": {
            "name": "Antonio Paya Gonzalez",
            "url": "http://antoniopg.tk",
            "email": "antonioalfa22@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/antonioalfa22/bms-monitoring/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/dashboard/finding-type-count/{id}": {
            "get": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "Get finding type count by Project ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Findings By Project Dashboard",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http_res.HTTPResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http_res.HTTPResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "data_count": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}